requirements.txt
python-dotenv
azure-cognitiveservices-speech
pydub
deepgram-sdk
groq
pydantic